+++ frontend/src/features/map_planning/components/Map.tsx
@@ -144,6 +144,7 @@ export const Map = ({ layers }: MapProps) => {
                   onClick={() => undo()}
                   title={t('undoRedo:undo_tooltip')}
                   data-tourid="undo"
+                  data-testid="undo-button"
                 >
                   <UndoIcon></UndoIcon>
                 </IconButton>
@@ -151,6 +152,7 @@ export const Map = ({ layers }: MapProps) => {
                   className="m-2 h-8 w-8 border border-neutral-500 p-1"
                   onClick={() => redo()}
                   title={t('undoRedo:redo_tooltip')}
+                  data-testid="redo-button"
                 >
                   <RedoIcon></RedoIcon>
                 </IconButton>